requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-cohere
langchain-groq
pinecone
python-dotenv
pypdf
streamlit
